+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## next
 
+- breaking: The concept of `Subject` has been renamed `VerbObject`, because that is
+  actually what it is. The subject is always the user. if you are using the
+  TypeScript types `Subject*` you need to use `VerbObject` instead
 - build: move types into index.ts to avoid special handling afterwards
 
 ## 0.8.0